Refactored and moved methods out of TermsLookupRepository.
